refactor(translate): 重构翻译功能，提取翻译服务为独立模块
将翻译相关逻辑从ApiService中提取到独立的TranslateService模块
简化组件中翻译功能的调用方式，移除重复代码
