+++ src/renderer/src/components/Popups/TextEditPopup.tsx
@@ -1,9 +1,7 @@
 import { LoadingOutlined } from '@ant-design/icons'
 import { loggerService } from '@logger'
-import { useDefaultModel } from '@renderer/hooks/useAssistant'
 import { useSettings } from '@renderer/hooks/useSettings'
-import { fetchTranslate } from '@renderer/services/ApiService'
-import { getDefaultTranslateAssistant } from '@renderer/services/AssistantService'
+import { translateText } from '@renderer/services/TranslateService'
 import { getLanguageByLangcode } from '@renderer/utils/translate'
 import { Modal, ModalProps } from 'antd'
 import TextArea from 'antd/es/input/TextArea'
@@ -43,7 +41,6 @@ const PopupContainer: React.FC<Props> = ({
   const [textValue, setTextValue] = useState(text)
   const [isTranslating, setIsTranslating] = useState(false)
   const textareaRef = useRef<TextAreaRef>(null)
-  const { translateModel } = useDefaultModel()
   const { targetLanguage, showTranslateConfirm } = useSettings()
   const isMounted = useRef(true)
 
@@ -103,21 +100,12 @@ const PopupContainer: React.FC<Props> = ({
       if (!confirmed) return
     }
 
-    if (!translateModel) {
-      window.message.error({
-        content: t('translate.error.not_configured'),
-        key: 'translate-message'
-      })
-      return
-    }
-
     if (isMounted.current) {
       setIsTranslating(true)
     }
 
     try {
-      const assistant = getDefaultTranslateAssistant(getLanguageByLangcode(targetLanguage), textValue)
-      const translatedText = await fetchTranslate({ content: textValue, assistant })
+      const translatedText = await translateText(textValue, getLanguageByLangcode(targetLanguage))
       if (isMounted.current) {
         setTextValue(translatedText)
       }
