+++ src/renderer/src/components/TranslateButton.tsx
@@ -1,9 +1,7 @@
 import { LoadingOutlined } from '@ant-design/icons'
 import { loggerService } from '@logger'
-import { useDefaultModel } from '@renderer/hooks/useAssistant'
 import { useSettings } from '@renderer/hooks/useSettings'
-import { fetchTranslate } from '@renderer/services/ApiService'
-import { getDefaultTranslateAssistant } from '@renderer/services/AssistantService'
+import { translateText } from '@renderer/services/TranslateService'
 import { getLanguageByLangcode } from '@renderer/utils/translate'
 import { Button, Tooltip } from 'antd'
 import { Languages } from 'lucide-react'
@@ -23,7 +21,6 @@ const logger = loggerService.withContext('TranslateButton')
 
 const TranslateButton: FC<Props> = ({ text, onTranslated, disabled, style, isLoading }) => {
   const { t } = useTranslation()
-  const { translateModel } = useDefaultModel()
   const [isTranslating, setIsTranslating] = useState(false)
   const { targetLanguage, showTranslateConfirm } = useSettings()
 
@@ -45,21 +42,12 @@ const TranslateButton: FC<Props> = ({ text, onTranslated, disabled, style, isLoa
       return
     }
 
-    if (!translateModel) {
-      window.message.error({
-        content: t('translate.error.not_configured'),
-        key: 'translate-message'
-      })
-      return
-    }
-
     // 先复制原文到剪贴板
     await navigator.clipboard.writeText(text)
 
     setIsTranslating(true)
     try {
-      const assistant = getDefaultTranslateAssistant(getLanguageByLangcode(targetLanguage), text)
-      const translatedText = await fetchTranslate({ content: text, assistant })
+      const translatedText = await translateText(text, getLanguageByLangcode(targetLanguage))
       onTranslated(translatedText)
     } catch (error) {
       logger.error('Translation failed:', error as Error)
